Split legacy graph library validation tests
